Fixed edgecases with histogram
Histogram now stores ranges using doubles. This eliminates any
rounding errors from integer arithmetic. Consistently passes test
cases.

To-Do:
Normalize histogram
Automate testcases --command line utility
Finish makefile
Documentation

+++ Histogram.hpp
@@ -14,8 +14,9 @@ public:
 
 private:
     void fillBins();
-    int m_binSize;
+    double m_binSize;
     int m_maxBins;
+
     std::vector<int> m_bins;
     std::vector<int> m_values;
     std::vector<int> m_occurances;
